Move FrequentInquiries to shared components; Apply July 2026 FAQ redline and link its 'draft a letter' CTA
